scripts/requirements.txt
loguru==0.5.3
Flask_Cors==3.0.10
neo4j==4.3.1
environs==9.3.2
scipy==1.5.2
ortools==9.0.9048
matplotlib==3.3.2
pandas==1.1.3
Flask==1.1.2
numpy==1.19.2
protobuf==3.17.3
websocket_client==1.1.0